+++ src/main/java/com/project/ddm/repository/StationRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface StationRepository extends JpaRepository<Station, Integer> {
+public interface StationRepository extends JpaRepository<Station, Long> {
     List<Station> findAll();
 }
